+++ src/main/java/uk/q3c/krail/core/option/KrailOptionModule.java
@@ -30,7 +30,12 @@ public class KrailOptionModule extends OptionModule {
     @Override
     protected void bindPermissionVerifier() {
         bind(OptionPermissionVerifier.class).to(KrailOptionPermissionVerifier.class);
+        bindVaadinOptionSource();
         bindOptionPopup();
+    }
+
+    protected void bindVaadinOptionSource() {
+        bind(VaadinOptionSource.class).to(DefaultVaadinOptionSource.class);
     }