+++ src/main/java/uk/q3c/krail/core/sysadmin/option/SelectedOptionSourcePanel.java
@@ -16,9 +16,9 @@ package uk.q3c.krail.core.sysadmin.option;
 import com.google.inject.Inject;
 import uk.q3c.krail.core.i18n.I18N;
 import uk.q3c.krail.core.option.OptionPopup;
+import uk.q3c.krail.core.option.VaadinOptionSource;
 import uk.q3c.krail.i18n.Translate;
 import uk.q3c.krail.option.Option;
-import uk.q3c.krail.option.persist.OptionSource;
 import uk.q3c.krail.persist.PersistenceInfo;
 
 import java.lang.annotation.Annotation;
@@ -36,7 +36,7 @@ public class SelectedOptionSourcePanel extends SourcePanel {
     private Class<? extends Annotation> selectedSource;
 
     @Inject
-    protected SelectedOptionSourcePanel(Translate translate, OptionSource optionSource, Option option, OptionPopup optionPopup) {
+    protected SelectedOptionSourcePanel(Translate translate, VaadinOptionSource optionSource, Option option, OptionPopup optionPopup) {
         super(translate, optionSource, option, optionPopup);
     }